Conflicts:
	AABB_tree/include/CGAL/internal/AABB_tree/AABB_drawing_traits.h
	CGAL_ImageIO/src/CGAL_ImageIO/Image_3.cpp
	Installation/cmake/modules/CGAL_Macros.cmake
	Installation/include/CGAL/gl.h
	Polyhedron/demo/Polyhedron/CGAL_demo/Viewer_interface.h
	Polyhedron/demo/Polyhedron/CMakeLists.txt
	Polyhedron/demo/Polyhedron/GlSplat/Shader.h
	Polyhedron/demo/Polyhedron/MainWindow.cpp
	Polyhedron/demo/Polyhedron/MainWindow.h
	Polyhedron/demo/Polyhedron/Plugins/AABB_tree/Cut_plugin.cpp
	Polyhedron/demo/Polyhedron/Plugins/Operations_on_polyhedra/Scene_combinatorial_map_item.cpp
	Polyhedron/demo/Polyhedron/Plugins/Surface_mesh/Scene_polyhedron_shortest_path_item.cpp
	Polyhedron/demo/Polyhedron/Plugins/Surface_modeling/Deform_mesh.ui
	Polyhedron/demo/Polyhedron/Plugins/Surface_modeling/Scene_edit_polyhedron_item.cpp
	Polyhedron/demo/Polyhedron/Plugins/Surface_modeling/Scene_edit_polyhedron_item.h
	Polyhedron/demo/Polyhedron/Polyhedron_demo_mesh_3_plugin_cgal_code.cpp
	Polyhedron/demo/Polyhedron/Polyhedron_demo_scale_space_reconstruction_plugin.cpp
	Polyhedron/demo/Polyhedron/Scene.cpp
	Polyhedron/demo/Polyhedron/Scene.h
	Polyhedron/demo/Polyhedron/Scene_draw_interface.h
	Polyhedron/demo/Polyhedron/Scene_implicit_function_item.cpp
	Polyhedron/demo/Polyhedron/Scene_item.cpp
	Polyhedron/demo/Polyhedron/Scene_item.h
	Polyhedron/demo/Polyhedron/Scene_nef_polyhedron_item.cpp
	Polyhedron/demo/Polyhedron/Scene_points_with_normal_item.cpp
	Polyhedron/demo/Polyhedron/Scene_points_with_normal_item.h
	Polyhedron/demo/Polyhedron/Scene_polygon_soup_item.cpp
	Polyhedron/demo/Polyhedron/Scene_polyhedron_item.cpp
	Polyhedron/demo/Polyhedron/Scene_polyhedron_item_k_ring_selection.h
	Polyhedron/demo/Polyhedron/Scene_polyhedron_selection_item.cpp
	Polyhedron/demo/Polyhedron/Scene_polyhedron_selection_item.h
	Polyhedron/demo/Polyhedron/Scene_polylines_item.cpp
	Polyhedron/demo/Polyhedron/Scene_textured_polyhedron_item.cpp
	Polyhedron/demo/Polyhedron/Viewer.cpp
	Polyhedron/demo/Polyhedron/Viewer.h
	Polyhedron/demo/Polyhedron/include/CGAL/gl_render.h
	Polyhedron/demo/Polyhedron/include/Point_set_3.h

+++ Polyhedron/demo/Polyhedron/Plugins/Operations_on_polyhedra/Scene_combinatorial_map_item.h
@@ -88,10 +88,10 @@ private:
        NbOfVbos = Facets_normals +1
    };
 
-   mutable std::vector<double> positions_lines;
-   mutable std::vector<double> positions_points;
-   mutable std::vector<double> positions_facets;
-   mutable std::vector<double> normals;
+   mutable std::vector<float> positions_lines;
+   mutable std::vector<float> positions_points;
+   mutable std::vector<float> positions_facets;
+   mutable std::vector<float> normals;
    mutable std::size_t nb_lines;
    mutable std::size_t nb_points;
    mutable std::size_t nb_facets;
